refactor(): bulk refactoring of users/organizations flow
